reorganized as a python package
